Update file permissions and refactor address derivation imports
This commit updates the file permissions for several files in the WailsJS runtime and modifies the imports for address derivation in the accountsHandler, coins, and useCreateVaultMutation files. The changes ensure consistency in the import paths and improve the overall structure of the codebase.
